+++ test_suite/tests/ui/string/validate/len_utf16_min_vs_len_utf16_max.rs
@@ -1,0 +1,6 @@
+use nutype::nutype;
+
+#[nutype(validate(len_utf16_min = 127, len_utf16_max = 63))]
+pub struct JsText(String);
+
+fn main () {}